リファクタリング VLM
