+++ Utils/Counter.cs
@@ -1,0 +1,32 @@
+public class Counter
+{
+    private int _count;
+    private readonly object _lock = new object();
+
+    public int Value
+    {
+        get
+        {
+            lock (_lock)
+            {
+                return _count;
+            }
+        }
+    }
+
+    public void Increment()
+    {
+        lock (_lock)
+        {
+            _count++;
+        }
+    }
+
+    public void Reset()
+    {
+        lock (_lock)
+        {
+            _count = 0;
+        }
+    }
+}
